reimplement load
The load method attempts to load the OCFL object at this.path when the
class is instantiated. If it finds an object at the path it will attempt
to read the top level inventory.json to create the current state of the
object (stored as this.currentState) in addition to reading each version
inventory and creating the state of that version as this.versions.

Helper methods getCurrentState and getVersions return the current state
and versions respectively.