Totally rethink ocflObject and associated tests
API has been totally reworked to operate around a single update method
that knows hows to create a new object or update and existing one. The
method can either be passed a source directory to use for the object
content or a function that will write content (with the deposit target
being passed to the function.

The class has a public API and a private one (methods beginning with __
are considered private and should not be used by calling class users).

Extensive usage examples are visible in ocflObject.spec.js.

Original code has been commented out for now - likewise in test script.