make deposit and backup paths single folder
If we use a path (subtree) relative to deposit or backup then cleanup
becomes really difficult as we have to walk back up the tree in order to
remove the path. So, if an id is passed in, just use that without
converting to a path. Likewise if a path is passed in - make it string
by replacing / with ''.